fuzz: build fuzz binary in Docker build steps
Running go-fuzz-build when the container launched requied +400MB and
was being OOMKilled on the node. Let's just bundle it with the image ourselves.

The image is a bit larger, but it was already big to begin with.

moov/achfuzz                v1.0.0              f570869bd778        31 minutes ago      918MB
moov/achfuzz                v0.6.0              a066990d34e5        2 hours ago         872MB